+++ test/link_checker_api_test.rb
@@ -41,22 +41,4 @@ describe GdsApi::LinkCheckerApi do
       assert_equal "http://example.com", batch_report.links[0].uri
     end
   end
-
-  describe "#upsert_resource_monitor" do
-    it "returns a useful response" do
-      stub_link_checker_api_upsert_resource_monitor(
-        reference: "Test:10",
-        app: "testing",
-        links: ["http://example.com"],
-      )
-
-      resource_monitor = @api.upsert_resource_monitor(
-        ["http://example.com"],
-        "testing",
-        "Test:10",
-      )
-
-      assert resource_monitor.key?("id")
-    end
-  end
 end